+++ src/components/user-signin/login/index.js
@@ -3,13 +3,13 @@ import { setUser, resetUser } from 'modules/user/actions';
 
 import Component from './component';
 
-const mapStateToProps = state => ({
-  user: state.user
+const mapStateToProps = (state) => ({
+  user: state.user,
 });
 
 const mapDispatchToProps = {
   setUser,
-  resetUser
+  resetUser,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Component);
