+++ src/components/user-signin/register/index.js
