+++ src/pages/home/component.js
@@ -6,10 +6,12 @@ import Section3 from 'components/home/section3';
 
 import './styles.scss';
 
-const Home = () => (<div className="l-home">
-  <Section1 />
-  <Section2 />
-  <Section3 />
-</div>);
+const Home = () => (
+  <div className="l-home">
+    <Section1 />
+    <Section2 />
+    <Section3 />
+  </div>
+);
 
 export default Home;
