+++ src/pages/index.js
@@ -3,7 +3,4 @@ import Component from './component';
 
 const stateToProps = ({ router }) => ({ router });
 
-export default connect(
-  stateToProps,
-  null
-)(Component);
+export default connect(stateToProps, null)(Component);
