+++ src/pages/recover/styles.scss
@@ -1,0 +1,8 @@
+@import 'styles/settings';
+
+.l-recover {
+  background-position-x: center;
+  background-position-y: top;
+  background-repeat: no-repeat;
+  height: 100vh;
+}
